refactor: store-request 개별 폴더링 (store로부터 출가하였음)
