+++ src/stores/stores.controller.ts
@@ -5,60 +5,11 @@ import { ReadStoreDTO } from './DTO/read-store.dto'
 import { ReadStoreAddressDTO } from './DTO/read-store-address.dto'
 import { StoreCategory } from './entities/store-category.enum'
 import { UpdateStoreDetailDTO } from './DTO/update-store-detail.dto'
-import { CreateStoreRequestDTO } from './DTO/create-store-request.dto'
-import { ReadStoreRequestDTO } from './DTO/read-store-request.dto'
-import { UpdateStoreRequestDTO } from './DTO/update-store-request.dto'
 
 @Controller('api/stores')
 export class StoresController {
     constructor(private storesService: StoresService) { }
 
-
-    // --------------------
-    // Store Request 가게 신청서
-
-    // CREATE
-    // 가게 신청서 생성
-    @Post('/request/')
-    async createStoreRequest(@Body() createStoreRequestDTO: CreateStoreRequestDTO): Promise<void> {
-        await this.storesService.createStoreRequest(createStoreRequestDTO)
-    }
-
-    // READ
-    // 모든 가게 신청서 조회 (관리자 전용)
-    @Get('/request/')
-    async getAllStoreRequest(): Promise<ReadStoreRequestDTO[]> {
-        const storeRequests = await this.storesService.getAllStoreRequest()
-        if (!storeRequests) { throw new Error('No storeRequests found') }
-
-        const readStoreRequestDTO = storeRequests.map(storeRequest => new ReadStoreRequestDTO(storeRequest))
-        return readStoreRequestDTO
-    }
-
-    // 특정 가게 신청서 조회
-    @Get('/request/:request_id')
-    async getStoreRequestById(@Param('request_id') request_id: number): Promise<ReadStoreRequestDTO> {
-        const storeRequest = await this.storesService.getStoreRequestById(request_id)
-        if (!storeRequest) { throw new Error(`StoreRequest with ID ${request_id} not found`)}
-        
-        return new ReadStoreRequestDTO(storeRequest)
-    }
-
-    // UPDATE
-    // 가게 신청서 처리 (관리자 전용)
-    @Put('/request/:request_id')
-    async updateStoreRequest(@Param('request_id') request_id: number, @Body() updateStoreRequestDTO: UpdateStoreRequestDTO) {
-        await this.storesService.updateStoreRequest(request_id, updateStoreRequestDTO)
-    }
-
-    // DELETE
-    // 가게 신청서 삭제
-    @Delete('/request/:request_id')
-    async deleteStoreRequest(@Param('request_id') request_id: number) {
-        await this.storesService.deleteStoreRequest(request_id)
-    }
-
-    // --------------------
     // CREATE
     // 새로운 가게 생성하기
     @Post('/')
