+++ src/stores/stores.module.ts
@@ -3,11 +3,10 @@ import { TypeOrmModule } from '@nestjs/typeorm'
 import { StoresService } from './stores.service'
 import { StoresController } from './stores.controller'
 import { Store } from './entities/store.entity'
-import { StoreRequest } from './entities/store-request.entity'
 
 @Module({
     imports: [
-        TypeOrmModule.forFeature([Store, StoreRequest])
+        TypeOrmModule.forFeature([Store])
     ],
     providers: [StoresService],
     controllers: [StoresController],
